Update solver.rb
Burke & Bykov (2016)  specify that 'beats_current_solution' should be true when it is better (less in case of minimization)  or EQUAL to the current solution.  This change may have a significant impact when the objective function of the solution is discrete-valued, or if there are many 'plateaus'  in the function to be optimized. It will accept equal-valued new solutions, thus allowing for more diversification.